+++ src/main/java/com/ureca/juksoon/domain/feed/dto/response/CreateFeedRes.java
@@ -1,4 +1,4 @@
-package com.ureca.juksoon.domain.feed.dto.responce;
+package com.ureca.juksoon.domain.feed.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
